Zobrist hashes
This adds Zobrist hashes to the intset class.  Zobrist hashes have the
beneficial property that they can be calculated as elements are added
and removed from the set: each possible element of the set has its own
randomly generated key, which is XORed into the hash value each time
that element is added or removed.