Using judyhash library
It turns out there's already a library (judyhash) that provides a C++
interface into the Judy libraries, including building a proper
polymorphic hash table class.  Which is nice since now I don't have to
create one myself.  This patch modifies the intset code to use the
judyhash library.

I'm including the judyhash library directly in the source code, since
it is small, only consists of header files, and is very unlikely to be
available in any current packaging systems.  Also, there are currently
a couple of minor bugs with the Makefile's install target, so it's
easiest just to drop a copy inside our source and include the files
directly from there.  Of course, this means that I'm now responsible
for installing the judyhash include files; so once the Makefile issues
are resolved, I'll probably need to remove it as an internal library.