Eventmaps
An eventmap is a relation (in the mathematical sense) between events.
They will be used by the renaming and linking CSP operators; for
instance, in the case of renaming, the eventmap specifies which events
should be renamed, and which new names they should get.  Currently
there are basic operations for:

  • adding and removing pairs of events
  • testing for membership of a certain pair
  • counting the number of pairs in the eventmap
  • retrieving all of the pairs into a preallocated array

Some basic test cases are included; more will probably be needed.