Initial CSP representation; operational semantics for prefix
This initial representation of a CSP script is a fairly simple wrapper
around an LTS.  Processes are represented as states.  Processes and
events can be named, which are stored in respective symbol tables.

Currently we've only added the operational semantics for the prefix
operator.  Others will come later.

This patch also introduces the beginnings of the CSP0 language (again,
with initial support only for naming processes and events, and the
prefix operator).  A simple test harness is included that reads CSP0
scripts and outputs the corresponding internal CSP representation.