Fixed null pointer dereference when examining LTS edges
Several of the methods for examining the edges in an LTS would
dereference a shared_ptr without first verifying that it was not null.
The csp/00-extchoice3 test case highlights this incorrect behavior,
since it must walk the non-existent outgoing edges from the STOP
process.  These methods now all correctly check for NULL pointers.