Memoization of CSP processes
There are now two methods for each CSP operator.  The existing one
required the process's LTS state to have already been created.  This
mechanism will be needed to define recursive processes, since at some
point a recursive process will have to refer back to a process that
has been created but not finalized.

The new method creates the LTS state (using the add_temp_process()
method) and adds the appropriate edges in one call.  Moreover, the new
method is memoized, so that if we've already created a process using
the same parameters, it is reused rather than recalculated.  This is
especially helpful with the interleave operator, since many of the
states in the interleave's expansion can be reused.