Using Bison for all parsing routines
Bison is now used to implement all of the parsing routines.  We don't
use Flex for the scanner, since the Flex buffering routines made it
difficult to scan only part of an input stream.  Since our grammar is
relatively simple, it was easier to write a custom lexer that uses the
underlying buffering of the istream classes.

As a result of this change, the alphabets in a CSP0 script should now
contain event names instead of event numbers, which was the originally
desired behavior.