Event closure
As the first part of implementating the normalization algorithm for
LTSs, I've written the code to find the closure of a set of states for
a particular event.  The closure is the set of nodes that can be
reached from any of the initial nodes by following any sequence of
that event.  When we get to CSP normalization, we'll be looking
specifically for τ closures.

The pseudo-code algorithm is:

  CLOSURE(initial_nodes, event): closure_nodes
    seen = ∅
    closure_nodes = ∅
    queue = initial_nodes
    for x ∈ queue:
      seen ∪= {x}
      closure_nodes ∪= {x}
      queue ∖= {x}
      for y ∈ afters(x, event):
        if y ∉ seen:
          queue ∪= {y}