Refactoring LTS graph into separate class
This patch extracts the LTS's inner event→stateset map into a separate
class.  This will allow us to use this map class on its own when we
prenormalize an LTS.  As part of this, I've also gone through and made
sure that every time we declare a shared_ptr to a class, we also
declare a shared_ptr to a const version of the class.