Initial support for LTS prenormalization
This patch adds support for prenormalizing an LTS.  The algorithm used
is described in [Roscoe94, §2.1, Stage 1].  The nodes in a
prenormalized LTS each represent a set of states from the original
LTS, and each node is guaranteed to have at most one outgoing
transition for each event.

The prenormalization algorithm isn't complete yet, however, since it
only supports the traces model.

  [Roscoe94] A. W. Roscoe.  “Model-checking CSP”.  In A. W. Roscoe,
     editor, /A classical mind: Essays in honour of C. A. R. Hoare/,
     pages 353–378.  Prentice-Hall, 1994.