Equivalence classes
This patch adds support for equivalence classes.  An equivalence
class is a set of sets, where each member of an inner set are
thought to be “equal” in some way.  Methods are available for
adding states to an equivalence class and for merging existing
equivalence classes.  Test cases are provided.