Traces normalization
This patch adds support for normalizing LTS.  The algorithm used is
described in [Roscoe94, §2.1], and combines the prenormalization and
bisimulation code from previous patches.  Normalized LTSs are now
constructed in stages; you can prenormalize as many initial source
states as you wish, but once you finalize the normalization, no more
prenormalizations are allowed.

As with previous patches, this algorithm currently only works with the
traces model.

  [Roscoe94] A. W. Roscoe.  “Model-checking CSP”.  In A. W. Roscoe,
     editor, /A classical mind: Essays in honour of C. A. R. Hoare/,
     pages 353–378.  Prentice-Hall, 1994.