Fixed recursion and memoization support
The CSP operator implementations were broken when it came to
recursion.  This was easiest to see with interleaving (see the
tests/csp/05-rec-interleave1 test case), which would go into an
infinite loop when you tried to compile it.  I've fixed this for all
of the operators (with test cases).  As a result, I also fixed the
memoization code, so that both versions of each operator are memoized.
(In the case where a process has already been allocated for the
operation, I've implemented this a single τ transition to the memoized
version.)