Fixed strict-aliasing warning in release builds
The release builds, which turn on compiler optimatizations, generate a
strict aliasing warning when compiling the parser files.  This is
because the scanner is written in terms of a “least common
denominator” Bison semantic type.  We are careful to define all of the
Bison semantic types so that they begin with the same elements as the
one used by the scanner.  However, the typecase that we use when
calling the lexer violates the strict aliasing constraints, which
don't allow you to perform type punning using pointer casts.  To get
around this, I've turned off the strict aliasing optimizations for
these compilation units.