Renaming TAU and TICK so that they are hidden
Previously, the τ and ✓ methods were given the names “TAU” and “TICK”
when created during the initialization of a csp_t object.  However,
these names mean that the events were unintentionally visible to a
CSP₀ script.  This made it possible to, for instance, use a prefix
statement to explicitly add a τ or ✓ to a CSP LTS, instead of only
being able to introduce them using the CSP operators.  I've renamed
them to “%TAU” and “%TICK”; since the percent character isn't a valid
identifier character in CSP₀, they are now correctly hidden in the
script.  It's now possible to define your own events called “TAU” and
“TICK”; however, these user-defined events won't actually have the
semantics of τ and ✓, so you shouldn't do that.