Adding acceptance sets for LTSs
This patch adds support for acceptance sets in an LTS.  Each stable
node in an LTS (i.e., those with no outgoing τ edges) now has an
acceptance set, which is the complement of the refusal set more often
seen in the literature.  If a state has some acceptance set that does
not contain event ‘a’, then there is some situation in which the state
can refuse to perform ‘a’.  These acceptance sets are needed to
implement the failures and failures-divergences semantic models in
CSP.

In this patch, LTSs can now store acceptance information.  As with LTS
edges, acceptances can only be added to a state that has not yet been
finalized.  The CSP operators add acceptances to the underlying LTS.
However, the normalization and refinement checking code does not yet
do anything with this information.

Lighthouse: [#3]