Interleave now synchronizes on ✓
The ✓ event must always be part of the synchronization alphabet for
any parallel composition, including interleaving.  The interface
parallel operator was already correctly synchronizing on ✓, so I've
refactored the interleaving code to simply call the interface parallel
code with an empty synchronization alphabet.  (This is similar to
alphabetized parallel, which is also implemented as a call to
interface parallel.)  This eliminates some redundant code, and ensures
that all of the parallel composition operators will now have the same
behavior.

Lighthouse: [#6 state:resolved]