Failures normalization
With this patch, the normalization code now works in the failures
model as well as the traces model.  The normalization code now takes
as a parameter which semantic model to use.  Currently, failures and
failures-divergences are treated the same, and do not consider
divergence information.  The tests/csp/80-roscoe-02 test case is one
example where the normalization is different in the two models.

The interface to this code could probably use some cleaning up, but
it's also probably not worth doing this until the failures-divergences
code is added.

Lighthouse: [#3]