Replicated external choice
This patch introduces the “replicated external choice” operator.  This
has the same semantics as the standard external choice, but it works
on a set containing any number of processes.  (The standard external
choice operator takes exactly two operands.)  The replicated version
is well-defined since the binary version is associative; this means
that it doesn't matter which order we apply the binary version to the
elements of the set.  Since the unit of binary external choice is
STOP, we also get STOP if we apply the replicated version to the empty
set.

Lighthouse: [#8]