Initial Haskell CSPM implementation
This is an initial patch implementing the CSPM language in Haskell.
I've chosen Haskell since, like CSPM, it's a lazy functional language.
This means we don't have to “implement” any of the functional features
of CSPM; rather, we can just call the appropriate Haskell functions.

This patch includes a basic implementation of the simple non-process
expressions in CSPM.  It does not yet include variable references,
‘let’ bindings, or functions, since they'll all require static-scoped
environments.  There is a basic test suite, based on QuickCheck.  I
also still need to wrap everything up in a Cabal package.

Lighthouse: [#2]