CSPM environments
An environment is the symbol table that hold CSPM definitions.  They
provide a standard mapping between identifiers and the expressions
that define them.  Environments can also have an optional parent,
which allows nested scoping; if an identifier is not found in the
current environment, we then proceed to look for it in the containing
environment.

This patch introduces the basic API for creating and using
environments, but does not extend the evaluation rules to use them.

Lighthouse: [#11]