Flattening the Expression type
Before, we had separate datatypes for each “class” of expression; for
instance, a Number type for any expression that always evaluates to a
number.  The Expression type was then defined as a sum type across all
of these subtypes.

This patch flattens things so that there's only a single Expression
type.  This simplifies things by removing a level of indirection in
the type definitions, and also will make it easier to define “bound
expressions” as we implement static-scoped environments.

Lighthouse: [#11]