New implementation of sets
This patch introduces a new implementation for sets.  It's based on
the Data.Set standard library, augmented to also allow open ranges.
This implementation gives a canonical representation for every set,
making equality comparisons easy.  Unfortunately, this representation
might not work out completely, since the only infinite sets that are
directly supported are open ranges.  The powerset of an open range,
for instance, causes a stack overflow, since we eventually try to
create a Data.Set instance of the result.  This requires reading in
the entire powerset, which, of course, borks.

Lighthouse: [#12]