Infinite sets
This patch introduces another new implementation for sets, which
supports any infinite set.  (The previous implementation only
supported open ranges.)  A set is now defined by a “loader list”.  The
loader list can be infinite, and can contain duplicates.  Sets can be
accessed in one of two ways.  The toSet function turns the set into an
instance of the standard Data.Se  t type.  As such, it only works on
finite sets.  The toList function, on the other hand, returns the
elements as a lazy list.  It therefore supports infinite sets.  As
elements are extracted from the loader list, they are added to a
“storage set”; the result is that the toList function returns the
original loader list, with duplicates removed.

Lighthouse: [#12 state:resolved]