Binding and evaluation now take place in a monad
Once we add process terms to the CSPM language, we'll need a State
monad to bind and evaluate them, since we'll need to generate new,
unique process names during evaluation.  This patch refactors the
existing binding and evaluation code to use a State monad.  Right now,
the state is empty, but once we introduce processes, it will contain
the list of used process names.

Lighthouse: [#2]