Distinct process names for each application of a function
Before, a process created by a function application would get its name
based on the context in which the lambda was defined.  This meant that
if you called a function many times, each with different values for
its parameters, you'd only get a single process definition in the
output script.  Ideally, we'd add to the process name some
representation of the values that the function was called with.
Unfortunately, that's not possible with the current design, since the
parameters to a function are lazy.  We bind the function application
(thereby creating its process name) before evaluating the lambda's
body, and we cannot evaluate any of the parameters before evaluating
the lambda's body.  Thus, we cannot use the parameter values to
calculate the process name.

We could probably reorganize things to get around this, but for now,
we just keep track of a global counter of function applications, and
append this counter to the process name created for any application.
This means that if we apply a function many times with the same value,
we will get distinct processes for each one.

Lighthouse: [#2]