Helper methods for evaluating CSPM expressions
This patch introduces two helper functions for evaluating CSPM
expressions: evaluateWith and evaluate.  They wrap together the act of
binding an expression, creating an Eval monad instance, and running
that instance.  evaluateWith allows you to specify which evaluator to
use, in case you need to coerce the value that results.  In most
cases, though, you can use evaluate as follows:

  evaluate env expr