Basic CSPM parser and command-line utility
This patch introduces a initial, basic CSPM parser.  It does not yet
support the entire CSPM language, or even the entire subset of the
language that our underlying evaluation engine can handle.  We can
handle most basic expressions, except for those operations which are
implemented as built-in functions in CSPM.

We've also created a “cspm” command-line utility, which parallels the
existing “csp0” utility.  Currently, it only supports one command —
“evaluate”.  This command loads in a CSPM script (either from a file
or specified on the command line), and evaluates a list of expressions
in terms of that script.  While a useful utility in its own right,
this will also serve as the basis for a CSPM evaluation test suite.