Top-level script no longer implemented as an implicit “let” statement
Before, we implemented the top level of a CSPM script the same way as
a “let” statement — as a list of Bindings.  This isn't exactly right,
though, since there are statements that can appear at the top level of
a script that aren't valid in a “let” statement.  In Scattergood's
grammar, this difference is encoded in the Defn and SimpleDefn rules.
This patch implements that separation.  This will allow us to start
implementing the other possible top-level statements, such as
“channel”.