Allow newlines before and after binary operators
The CSPM spec uses newlines to separate the definitions in a file and
in a “let” statement, but also allows newlines to appear certain extra
situations for readability — for instance, before and after any binary
operator.  Like in [Scattergood98], we handle these silent newlines in
the lexer rather than the parser; any consecutive newlines are
replaced with a single TNewline token, and any TNewlines that appear
before or after a “soakable” token are removed from the token stream
before it's passed on to the parser.

  [Scattergood98] J. B. Scattergood.  /The semantics and
    implementation of Machine-Readable CSP/.  D.Phil. dissertation,
    Oxford University, 1998.