Implementing built-in functions as keywords
This patch adds support in the CSPM compiler for the built-in CSPM
functions.  We've implemented these as keywords, instead of trying to
pre-seed a root environment with actual function definitions.  There
are test cases for all of the functions except for “Seq”, since it
returns an infinite set.  (We can define a “take” function that would
allow us to test, at least partially, infinite sequences.  But
currently we can't define an analogous function for sets.)