Initial pattern matching support
This pattern introduces some underlying functionality for implementing
pattern matching.  We've defined an ADT to represent patterns, and a
“valueMatches” function that determines whether a particular value
matches a pattern.  Most of the patterns defined by CSPM are included;
the two that are missing are channel names (which only match the
corresponding event), and dotted channel expressions, which we haven't
implemented yet at the expression level, either.

This patch also doesn't add pattern support to the CSPM parser or
expression evaluator.