Patterns on left-hand sides of top-level definitions
This patch allows arbitrary patterns to occur on the left-hand
side of a top-level definition.  Evaluation of the pattern match
is deferred until the usual evaluation phase.  This required a
new kind of Expression: a “pattern match extraction”.  This
expression evaluates the pattern match, and if it succeeds,
extracts the value of the specified identifier from the list of
bindings.  Old-style definitions, with a single identifier on the
left-hand side, are now special cases of a pattern definition.