“let”s now allow pattern definitions and multiple-branch functions
This patch extends the parser and AST evaluator to allow the same
kinds of definitions in a “let” statement as can appear at the top
level of a script.  Specifically, the left-hand side of a let binding
can now be an arbitrary pattern, or a function defined using multiple
branches.