“create-source-package” script now works with any version
The “create-source-package” script now takes the HST version as a
parameter, so it can be used to reproduce the tar archive for any
prior release.  It has also been updated to work with the latest Git
version, which no longer stores the (majority of the) git-* binaries
in /usr/bin.