Adding scripts for creating Mac OS X installation package
This patch introduces a script and some ancillary files for creating a
Mac OS X installation package.  We create an framework to hold the
CSP₀ library and command-line tool.  The installation package installs
this framework into /Libraries/Frameworks, and creates a symbolic link
in /usr/bin to point at the csp0 executable.

The package doesn't include the Haskell CSPM code yet; I still need to
figure out the best way to package it and its GHC dependency.