Including “cspm” in Mac OS X installation package
We now include the “cspm” Haskell program in the Mac OS X installation
package.  We still don't include the Haskell library itself in the
package, though.  Mostly this is because any library we include in the
package will (I believe) be tied to a specific compiler and version.
I'm thinking that anyone who needs the library can just grab it from
Hackage.  The “cspm” tool is statically linked, so we don't need the
library installed for it to work.  We shall see.