Updating CSPM libraries to work with GHC 6.10 and Cabal 1.6
The HST libraries do not use any of the new features or libraries
provided by GHC 6.10.  This patch ensures that we use the 6.8-era
base-3 libraries.  We also require the latest Cabal, which requires a
slight tweak to the Setup.lhs script to use the new name of the
default UserHooks instance.