Use shared_ptr from Boost instead of from the TR1 standard
When we first used shared_ptr in the HST code, we weren't using Boost
yet.  shared_ptr is in TR1, which seems to be implemented by most
modern C++ compilers, so it made sense to use that implementation,
rather than introduce a dependency on Boost.

Now, though, we're using Boost for other things, so it makes sense to
just use the shared_ptr implementation from it.  This also fixes a bug
reported by Moritz Kleine, where we'd get an error when trying to
compile src/assertions/refine.cc, since the compiler could somehow see
both implementations of shared_ptr, and would get confused which one
to use.

As part of this patch, I've also removed all “using namespace”
statements from the header files, since I've discovered that this is
rude behavior in the C++ world.  (We don't want to require the users
of the header files to use the namespace, too, after all...)