Replicated internal choice
This patch introduces the “replicated internal choice” operator.  This
has the same semantics as the standard internal choice, but it works
on a set containing any number of processes.  (The standard internal
choice operator takes exactly two operands.)  The replicated version
is well-defined since the binary version is associative.  Unlike
replicated external choice, the replicated internal choice of the
empty set is not well-defined.  Thus, the following is treated as a
syntax error in CSP₀:

  rintchoice P = |~| {};

Lighthouse: [#8]