Int and Bool built-in types
This patch introduces initial support for named types in CSPM.  Types
treated the same as sets; the only difference is in that certain
contexts, the tuple constructor and dot operator have different
semantics, which make it easier to define the kinds of sets that types
usually consist of.

The first two named types are the built-in Int and Bool types.  They
are defined as keywords, and have built-in evaluation rules that
result in the appropriate {True, False} and {..., -1, 0, 1, ...} sets.

Lighthouse: [#17]