“nametype” statements
The CSPM parser now supports “nametype” statements at the top level of
a CSPM script.  A nametype statement is a simple binding of a variable
to a type expression.  There is not yet support for the tuple operator
in type expressions; a later patch will introduce its product type
semantics.

Lighthouse: [#17]