Fixed column numbers when using Bison 2.3
Bison 2.3 numbers the columns in an input file from 0, whereas 2.4
numbers them from 1.  We want to consistently use 1-based column
numbers.  This was causing a problem with test cases that would
generate (expected) parse errors — if the correct test result was
generated using Bison 2.4, then that test would fail when using 2.3,
and vice versa.

I could not find an easy way to override the “column” field's default
value, so instead, we add code in our Bison grammars to manually reset
the column to 1 at the start of each file, and every time a new line
is encountered.

Lighthouse: [#20 state:resolved]