Simple datatypes
This patch introduces support for “simple” datatype statements — i.e.,
those whose constructors do not take any data values.  So, in the
following:

  datatype SimpleColour = Red | Green | Blue
  datatype T = A.{0..3} | B.{4..7}

SimpleColour is a simple datatype, whereas T is not.

The SimpleColour definition creates four bindings: Red, Green, and
Blue all resolve to a new “constructor value”, which just records
which of the data constructors was used to create a particular value.
SimpleColour resolves to a set containing all three constructor
values.

Lighthouse: [#21]