Eliminating CMake warnings
This patch fixes the CMake build scripts to eliminate two CMake
warnings.  Specifically, the names of each target must be unique;
before, there were several test cases in the different subdirectories
that had identical names.  They now have names that match the
corresponding test names.  The target names are used as-is as Makefile
targets, so we must be sure to only use characters that are valid in a
path.  The second warning is that we did not state a minimum CMake
version, which we now do.